WIP refact hour data parsing
