+++ src/adapters/opening_hour.js
@@ -1,3 +1,5 @@
+import I18n from '../libs/i18n'
+
 const days = ['mo', 'tu', 'we', 'th', 'fr', 'sa', 'su'];
 function parse(rawOpening) {
   var result = null;
@@ -72,4 +74,40 @@ function parse(rawOpening) {
   return result;
 }
 
-export default parse
+
+function openingStatus(openingData) {
+
+  if (openingData && openingData['24/7']) {
+    return 999
+    /* be sure it won't close soon */
+  }
+  let remoteDate = new Date()
+
+  if (!openingData) return -1
+  let dn = remoteDate.getDay()
+
+  let schedules = openingData[I18n.days[dn]]
+  if (!schedules) return -1
+  let open = schedules[0]
+  let close = schedules[1]
+
+  let currentTime = remoteDate.getHours() * 60 + remoteDate.getMinutes() //convert time to minutes
+
+  let [hours, minutes] = open.split(':') // time format is hh:mm
+  let openingTime = parseInt(hours) * 60 + parseInt(minutes);
+
+  [hours, minutes] = close.split(':')
+  let closingTime = parseInt(hours) * 60 + parseInt(minutes)
+  if (openingTime < closingTime) { // 10h00 14h30
+    if (currentTime > openingTime && currentTime < closingTime) {
+      return closingTime - currentTime
+    }
+  } else { // 17h00 2h00
+    if (currentTime < openingTime || currentTime > closingTime) {
+      return currentTime - closingTime
+    }
+  }
+  return -1 // closed
+}
+
+export {parse, openingStatus}
