+++ src/adapters/poi_popup.js
@@ -1,5 +1,8 @@
 import {Popup} from 'mapbox-gl--ENV'
+import {parse, openingStatus} from '../../src/adapters/opening_hour'
 import Poi from '../mapbox/poi'
+import IconManager from "./icon_manager";
+const poiSubClass = require('../mapbox/poi_subclass')
 let popupTemplate = require('../views/popup.dot')
 
 const WAIT_BEFORE_DISPLAY = 800
@@ -7,7 +10,7 @@ const WAIT_BEFORE_DISPLAY = 800
 
 function PoiPopup() {}
 
-PoiPopup.prototype.init = function(map, interactiveLayer) {
+PoiPopup.prototype.init = function(map) {
   this.map = map
   this.popupHandle = null
 
@@ -34,9 +37,12 @@ PoiPopup.prototype.addListener = function(layer) {
 
 PoiPopup.prototype.create = async function (layerPoi) {
   let poi = await Poi.apiLoad(layerPoi.properties.global_id)
-  this.popupHandle = new Popup({className: 'poi_popup'})
+  let className = poiSubClass(poi.subClassName)
+  let {color} = IconManager.get(poi.className, poi.subClassName)
+  let opening = openingStatus(parse(poi.opening))
+    this.popupHandle = new Popup({className: 'poi_popup'})
     .setLngLat(poi.getLngLat())
-    .setHTML(popupTemplate.call(poi))
+    .setHTML(popupTemplate.call({poi, className, color, opening}))
     .addTo(this.map)
 }
