+++ src/panel/poi_bloc/hour_panel.js
@@ -1,7 +1,6 @@
 import HourPanelView from '../../views/poi_bloc/hour.dot'
 import Panel from "../../libs/panel";
-import openingHourParse from '../../../src/adapters/opening_hour'
-import I18n from '../../libs/i18n'
+import {parse, openingStatus} from '../../../src/adapters/opening_hour'
 const strftime = require('strftime')
 
 function HourPanel(block, poi, options) {
@@ -9,54 +8,18 @@ function HourPanel(block, poi, options) {
   this.name = block.name
   this.timeMessages = options.messages
   this.title = options.title
-  this.hours = openingHourParse(block.raw)
+  this.hours = parse(block.raw)
   this.nextTransition = getSecondBeforeNextTransition(block)
   this.displayHours = translateHours(this.hours)
   this.latLng = poi.latLon
   this.status = this.computeStatus()
 }
 
-HourPanel.prototype.computeRemainingTime = function() {
-  let rawDate
-  if(this.hours && this.hours['24/7']) {
-    return 999 /* be sure it won't close soon */
-  }
-  rawDate = new Date()
-
-  let remoteDate = new Date(rawDate)
-
-  if(!this.hours) return -1
-  let dn = remoteDate.getDay()
-
-  let schedules = this.hours[I18n.days[dn]]
-  if(!schedules) return -1
-  let open = schedules[0]
-  let close = schedules[1]
-
-  let currentTime = remoteDate.getHours() * 60 + remoteDate.getMinutes() //convert time to minutes
-
-  let [hours, minutes] = open.split(':') // time format is hh:mm
-  let openingTime = parseInt(hours) * 60 + parseInt(minutes);
-
-  [hours, minutes] = close.split(':')
-  let closingTime = parseInt(hours) * 60 + parseInt(minutes)
-  if(openingTime < closingTime) { // 10h00 14h30
-    if(currentTime > openingTime && currentTime < closingTime) {
-      return closingTime - currentTime
-    }
-  } else { // 17h00 2h00
-    if(currentTime < openingTime || currentTime > closingTime) {
-      return currentTime - closingTime
-    }
-  }
-  return -1 // closed
-}
-
 HourPanel.prototype.computeStatus = function() {
   if(!this.hours) {
     return {msg : '', color : '#fff'}
   }
-  let remaining = this.computeRemainingTime()
+  let remaining = openingStatus(this.hours)
   if(remaining === -1) {
     return {msg : this.timeMessages.closed.msg, color : this.timeMessages.closed.color}
   }
