+++ src/scss/includes/map_theme.scss
@@ -8,3 +8,8 @@
   background-color: #54c07c;
   color: #fff;
 }
+
+.poi_popup {
+  border-radius: 6px;
+  border-left: 3px solid;
+}
